rasppleii/index.md: Alter markup a bit
Kramdown can make this markdown page look a lot like Ivan's original
with just a bit of help from CSS if you have the right classes.  The
biggest problems there are the <br> tags.  Ivan's page used them
heavily, and while they do make sense in a couple of places, they are
poor semantic representations of what you're trying to do.

Fortunately Kramdown lets you specify classes using {: … } syntax.  It's
nonstandard and the CommonMark folks aren't sure they want to support
it, but we're using kramdown and it does.

Regarding the other places, I opted to use the standard two-space line
terminator.  Kramdown doesn't seem to support \ at the end of the line
for <br> which is the CommonMark spec, and it wants \\ which sometimes
it interprets ambiguously as a literal \ character.  I may revisit this
problem later.