Make style.css more generic; move to root css dir
I've managed to pretty accurately reproduce the look of rasppleii's
index.html, but aside from some hard-coded colors and the use of tags
not supported by MarkDown/CommonMark, a lot of that CSS can probably be
reused for the A2SERVER page at least.  The A2CLOUD page has more
complex, more modern CSS, but I'll cross that bridge when I come to it.

In the meantime, I'm moving the CSS I wrote for the Raspple II to the
docuent root and re-structuring the Raspple II page-specific colors to
work assuming that some parent tag (body in this case) sets the class
"rasppleii".  I've modified my erb template to provide that.

The template is not yet in version control mainly because I would want
to put it where it will eventually live from the start and I sort of
need to decide how I'm going to ultimately handle those generation
tasks.  Probably Jekyll or something that looks a damned lot like it,
unless I can come up with some better idea.