style.css: Add unstyled em and text code
a2server-comands.md had a span of fine-print used like an aside, which
really is a form of emphasis, so I decided to mark it as such.  That let
me add the .fine-print class, but the resultant text was italicized.
So I added an override em.no-style.

Next, there's an embedded text file which should appear as plain
pre-formatted text marked as code.  I created .language-text for that,
which is displayed as a block so the background applies to the block
element.  Perhaps I should specify inline-block there to be most
HTMLically correct, but that would differ from the visual presentation
on ivan's page.