rasppleii/index.md: Use .uline
Slight cleanup to style.css to remove duplicated strong attributes.
They're only used once in rasppleii/index.md, so it's better this way.

Actually at some point we should minify the CSS for optimal acces, but
for the time being I like to be able to read what I'm developing.  ;)