a2server_lan.md: Match ivanx.com style
This one posed a pretty serious problem because of the way source links
are handled on Ivan's page.  The [ eBay ] stuff.  That's done using some
non-breaking spaces, which is both ugly and caused a whole bunch of
problems when I tried to turn it into a list.

Which actually doesn't work because once you replace the spaces with
ASCII spaces, you can get the result to style.  But kramdown doesn't
seem to like nesting lists very well, and once I got that working I
still couldn't do the one in the second paragraph because kramdown
cowardly refuses to violate the HTML spec for some reason.  ;)

Ultimately I had to fall back on <span> tags which aren't as
human-readable, but produce the correct output when run through the
kramdown process.

The way you're supposed to do something like this is to do it the way
the CommonMark spec document does: Use a second stage filter to generate
the desired markup and keep the source human-readable.  I'm not prepared
to do that right now.  So there's some HTML spans mixed in for now.