a2server_wifi.md: Match ivanx.com style
This basically completes the A2SERVER pages for the first pass (matching
the ivanx.com style) to the most reasonable extent that I care to do so.
I've actually done a little of step #2 fixing links, but didn't really
try very hard.  I just noticed a few that needed it and made the
requisite changes when I saw them.

I want to go and do the CSS for A2CLOUD first.  That's just one big file
as far as we're concerned here, and I'm not going to try to duplicate
wordpress spew verbatim, but I can make it look enough like the WP site
for our purposes.  I've got a little tidying up to do first, and then I
think I will actually upload the templates I'm using for kramdown, if
not the scripts to generate .html files just yet.