a2cloud/index.md: remove "a2cloud-" from IDs
Before, I was depending on kramdown's implicit ID generation for
headers, so "A2CLOUD: What you need" would become #a2cloud-what-you-need
automatically.  Since all of the posts are prefixed with "A2CLOUD: ",
probably none of them should be.

Of course to change that in the archive would be a content change, so
instead we'll explicitly set the IDs we want on those headings for now.
When the headings are renamed in the future, the explicit IDs all over
the index.md can be dropped.