a2c/index.md: Note about the sticky sidebar
Nutshell: I looked for a jQuery/framework-free solution for doing this,
and I did not find one.  In fact, what I found were mostly solutions
that were jQuery-based while claiming to be "Javascript".  There are
probably good reasons to use a web template like bootstrap 3 in the
future, and that'd include jQuery.  But that's then.  For now, I'm not
implementing this by dragging a massive pile of JS framework into the
party for one little block.  It just doesn't fit the roadmap at this
stage.

Basicaly, I don't know how to figure out the block's position or how far
you've scrolled down the page without using jQuery.  If someone reading
this can explain it or point to a jQueryless function that does the
needed bit of magic, I'll revisit this.