a2c/index.md: Finish editing content
Links are still a mess, but I've gone through the content at least.
I'll have a look through the link list later and do something with that.

Also, I'm not 100% happy with the CSS as-is.  Spacing still isn't right
with the <hr> tags between posts (thought I fixed that, but I must have
unfixed it at some point.)  Also, like many lipsum-generated WP themes,
this thing uses a base font of 14px and uses fonts all the way down to
12px.  That's damned near Apple // font resolution, people!  That's
designed to be looked at, not read!

The usual base font size on these pages is 16px, which is supposed to be
somewhat equivalent to 12pt.  The CSS presently keeps the various
rem-based sizes, but I'll probably toss those as the code isn't complex
enough to really require them.  I'll do some CSS tweaking probably
tomorrow night and then go back to figuring out what to do with the
links in the next couple of days.

The end is in sight!