a2c/index.md: Use embedded YouTube links
Given that we now have strayed from the pure Markdown for this page
source into including some HTML, let's use a (simplified) version of the
YouTube embedded video links.  I kept the markdownified versions in the
page as comments.  @IvanExpert can decide if he wants to revert back to
the Markdown versions, keep it like it is now, or remove the commented
Markdown.