style.css: Size and other style tweaks
The Twentytwelve WP theme uses a lot of empty whitespace coupled with
tiny fonts, which means the derived version did likewise.  After
limiting the page to the (common) 960px, which is done because 960 has a
lot of factors for grid layouts and displays nicely on 1024x768
monitors, it then cut off 80px worth of padding around the content and
left a huge gutter.  All so you could have layered pretty effects if you
wanted to.

Well, we didn't want to, really.  And by cutting out that padding and
making the columns fit closer to being 100%, it was easy to make the
whole thing bigger in terms of fonts.  This I did.

Of course, this still didn't totally fit the nav stuff properly, so I
cut off all the "A2CLOUD: " prefixes back in ef5fc85, which made the
whole thing look nicer.

The last style change was that since titles to posts aren't links, it
didn't make sense to color them as links.  But leaving them uncolored
made them white and that didn't look very good.  So I split the
difference and used an even blend of the two colors.  I could've gone
with any number of existing colors on the page, but I actually decided
in the process that I actually rather do like #bef and will be using
that color in the future for other things.  :)