build_site.sh: Script to generate our website
Here's the script I wrote to generate our website from Markdown in the
directory _site.  You should not serve out of _site because its contents
are subject to being rm -rf'd every time build_site.sh is re-run.

What I'd suggest doing is implement a command argument to the script
named "publish" or something.  If you provide that, and if $2 is
nonempty, I'd run an rsync command after the generation of _site
finishes to copy the contents of _site to the remote.  Use of rsync even
on the local filesystem is worthwhile because with the right arguments
rsync can prune things in dest not in src.  Of course you need to have
other arguments to avoid pruning files/ subdirectories at the moment and
yeah I don't need that yet so I haven't implemented it.  :)