a2c/index: All links moved, redo comments
The comments have been restructured to have titles of their own (h3)
with the bylines (formerly h3) moved to h4 just committed.

All in-file id links have been changed from [text][ref] format to inline
[text](#id) format.  This is not 100% ideal in places, but it's a good
change in most cases and it's easier to undo it for an annoying case
later than to add to the list of [ref]: url lines whose refs are just
arbitrary numbers in a file the size of A2CLOUD's single page.  A
rewording of the text (which won't happen for awhile) will probably
remove the need to do it anyway.

The numbered links are still numbered, at least where an obvious named
ref didn't come to mind.  Another pass.