a2c/index: Shortened Amazon URLs
Amazon URLs include a lot of tracking and session info that doesn't
necessarily need to be included to find the URL you want.  You can
actually cut a lot out.  For an Amazon product URL, your shortest option
without a URL shortener (Amazon's or other) is:

    http://amazon.com/dp/<ASIN>

…where <ASIN> is the Amazon product number or a book's ISBN number.

A search is slightly more complex

    http://www.amazon.com/s?url=search-alias%3D<category>&field-keywords=<key+words>

…where <category> can be things like "electronics" or "computers" and
<key+words> are the plus-separated list of words in the search box.  The
probably shortest category is "aps" which is the all-of-Amazon category.
I suggest narrowing the category you want to search in to where you want
it and look at URL Amazon gives you for what to use.

Of course you might want to add ?tag=<assoc> or &tag=<assoc> as
appropriate.  In this commit, almost all of the Amazon links had a tag
for Ivan.  Say thanks for A2CLOUD by using them to buy!  :)