a2c/index: Last of linkrefs converted
There are no more numerical linkrefs in A2CLOUD's index.md.  There are
in the page for Raspple II and all of the A2SERVER pages, but these
aren't a priority because there's not 90 of them.

The other link optimizations I've done to A2CLOUD for amazon, ebay, etc.
could be done for other pages, but the big one was A2CLOUD and it's
*done*.