blog.html,index.md: Rearrange files
The default Jekyll index.html is a liquid/html page that spits out a
blog based on things in _posts.  We aren't using that really, so we'll
take it out of the rotation by renaming it to blog.html, but failing to
give it a title tag in the YAML front matter.  What this does is cause
it not to generate a top-level navigation link.

rasppleii/index.md moves to index.md in its place, becoming the root
page of the Raspple II website.  On Ivan's server, that's the Apple II
Extravaganza, but as we've set up the GitHub site, Raspple II is the
unbrella.  We can adjust things as necessary.