_layout.scss: Add (lame) .current for header
The current node in the header is given a highlight class, so for now
we'll mark that by inverting the text/background colors.  Yes, that's
lame, and we probably should use something more, well, *more* for the
highlighting.  But again, this creates the highlight.  Fancy styles can
wait for structural stuff.