footer.html: Reduce generated whitespace
One of the side-effects of Liquid templates is that they generate a LOT
of extraneous whitespace, especially for loops.  A large loop can have
thousands of blank spaces and newlines.  Fixing that makes for ugly
liquid, but we can make the liquid readable (if ugly) and still have
clean output.