header.html: Categories for top nav + gen white
Pages no longer go into the top nav based on whether or not they have a
title setting.  Rather, they go into the top nav if they have setting
for category:.  Additionally, the current category page is highlighted.

In order to have pages be considered part of a category, there is the
YAML setting in-category, which is used if the page has no category.

Finally, doing this caused me to observe the really terrible whitespace
spew out of the Liquid generation.  So whitespace has been eaten through
the use of ugly Liquid tags as has been done elsewhere.  Not pretty but
it works.