js/*: Add some standard JavaScript
This is me, breaking down and moving to bootstrap-sass.  There's just
too much to do for me to reinvent the wheel constantly, and others have
already done the work.  Yes, this is overdone for what it is, but it
works and the major pieces are things people already have in their
browser cache.