Finish cleaning Jekyll defaults out
Okay, I was a bit hasty removing Jekyll's defaults because it broke some
stuff.  First, there were syntax highlighting rules (which we don't want
anyway, at least not as they were) that require the rest of the Jekyll
sass defaults.  I've already committed the fix for that separately.

Next, since we still use the Jekyll SVG icons for Twitter and GitHub (I
like them!), we needed to keep them around.  Defaults called them to be
filled with a a gray we were deleting, but I went with the text color
since that's what the gray in the old theme was for.

Next, I was able to re-delete the _base.scss and _layout.scss files
since we really no longer use them.  In main.scss, I re-deleted all of
the base stuff since bootstrap does that for us.