main.scss: Add _bootstrap-theme.scss
The way this works is that Bootstrap's sass code will set a bunch of
variables if they're not already set, so we need custom-variables to be
loaded before the theme.  Then we need to load our theme after so that
it generates override CSS.

_bootstrap-theme.scss: The contents of this file at present style the
Bootstrap navbar element to look like it isn't a navbar at all.  The
brand is written in our case using the VT323 font.  If someone wants to
do that as a logo that seems better suited to the Apple //, I'm all for
it, especially if you define it as SVG and I can make it scale.

The right-justified nav links appear to be just links, and currently
they're rather unadorned (presently too unadorned), but they do respond
to hover/focus and highlight nicely indicating their linkiness.  The
current one is highlighted nicely in a fashion similar to the Bootstrap
pill nav elements, but with substantially tweaked settings.

All of this is designed to center as best it can be centered.  More to
come in future, but this will do for tonight.