theme-test.html: Theme tester from bootswatch
Okay, there's a ton of themes for bootstrap v3, but they're mostly
written in less.  Of course bootstrap is moving to sass/scss as of v4,
has an official sass/scss port of v3, and Jekyll uses sass/scss.  The
writing is on the wall, with the scss variant of sass, the less vs. sass
holy war is mostly over and sass is the winner.

Of course that means we have to do our themes by hand.  Which would be
fine if we had a nice fully bootstrapped website to work from.  We
don't though.  So we'd either have to fix it as we go, or write a test
page that uses every bootstrap widget we're going to want to theme!

BootSwatch already did that to show off their free Bootstrap themes,
which are pretty cool actually.  So I cheated and swiped theirs.