theme-test.html: Fix styles, javascript
The modal now shows as it should.  To do this, we need to override some
styles.  Since they're local to the page and the overrides are short,
I've just stuck them into the page where they belong.  This isn't ideal
practice because the styles may have somewhat undefined behavior
depending on the browser, but we'll risk it here since the major modern
browsers do what we want and this page isn't intended to be front-facing.

There's also added some JavaScript to make the popovers and tooltips do
what you expect.

Finally, there's the addition of the source button.  The major reason
for the bs-component class added to the bootswatch demo page is so that
you can get a nice little "show me the source code for this" button (and
as an anchor for the custom modal style.)  Plus the code for the source
dialog was already included from yesterday's fiddling, so I might as
well make it work.