a2server/**: Remove source code from website
The source code should not be in the website.  Certainly the source code
to an old version shouldn't be.  Nuke it.  We'll fix this later,
replacing it with links to the current version on GitHub.  I've already
started looking into how to do this and I think it'll be pretty cool
actually.  :)