a2server_commands.md: Include remote file
a2server_commands.md used to use an iframe to include the file
a2server-help.txt.  It's one of the reasons why source code needed to be
in the website's directory.  Here's a first pass at not needing to do
that anymore.

Here we use a little bit of jQuery to load the file from GitHub. "Ahh,"
you say.  "But this does not work if the user has JavaScript disabled,"
you say.  And you are correct.  It fails utterly without JS.  But the
fact that the JS works means that I can now disable JS and implement a
<noscript> block that uses an iframe.

Then once that works I can actually use the awesome powah of all this
liquid and yaml and whatnot to replace the ugly embedded html with a
nice, clean liquid include that sets a yaml variable for what remote
file to include.  I bet we can even syntax highlight it.