main.scss,_user-styles.scss: Hide JS for non-JS
If you have an element that needs to be hidden if JS is disabled, you
can now hide it by adding `class="need-js"` to the HTML tag.  In
Markdown, you do it with `{: .need-js }` (but it's less likely you'll be
doing that I suppose.

One step closer to a clean link to source code.