rfile.html: Replacement for iframe include
Tried a number of solutions and this is the least annoying I could come
up with within GitHub's framework.  To use, insert into any HTML or
Markdown source:

{% include rfile.html url='<base url>' file='<filename>' %}

The reason for this ugliness is that you cannot define a liquid tag on
GitHub pages.  It might be possible to pass url='<full url>' instead and
determine the filename by use of liquid brute force.  And I might do
that, but I want to commit something that works, even if it's kind of a
mess.