**.html: container-fluid -> container
It seems like container-fluid the way to go, but it's really not.  A
design can still be responsive (and ours is for all that it isn't
actually done yet), and yet still specify not to use the whole thing
when the width of the page is excessive.  That's a non-fluid container
and that's what we need to be using.